services/auth-service/requirements.txt
flask==3.1.3
flask-sqlalchemy==3.1.1
pymysql==1.1.1
cryptography==46.0.6
werkzeug==3.1.6
pyjwt==2.12.0
authlib==1.3.1
requests==2.32.3
Flake8
bandit
